+++ frontend/components/DescriptionOverwrite/Description/style.js
@@ -3,7 +3,6 @@ import { themeConfig } from '@shopgate/pwa-common/helpers/config';
 
 const container = css({
   fontSize: 14,
-  marginTop: -15,
   userSelect: 'none',
 }).toString();