+++ frontend/components/PropertiesOverwrite/Properties/style.js
@@ -3,9 +3,14 @@ import { themeConfig } from '@shopgate/pwa-common/helpers/config';
 
 const content = css({
   fontSize: 14,
-  padding: `0 ${(themeConfig.variables.gap.small * 1.375)}px ${themeConfig.variables.gap.big}px`,
+  padding: '0 16px',
   marginBottom: 12,
 }).toString();
+
+const contentAccordion = css({
+  padding: '0 !important',
+  marginBottom: '0 !important',
+});
 
 const label = css({
   paddingBottom: (themeConfig.variables.gap.small * 1.5),
@@ -13,7 +18,7 @@ const label = css({
 
 const cell = css({
   maxWidth: 100,
-  padding: '2px 5px',
+  padding: '2px 5px 2px 0',
   overflowWrap: 'break-word',
 }).toString();
 
@@ -23,6 +28,7 @@ const table = css({
 
 export default {
   content,
+  contentAccordion,
   label,
   cell,
   table,