+++ frontend/components/ReviewsOverwrite/theme-gmd/Reviews/components/Header/components/WriteReviewLink/index.jsx
@@ -1,11 +1,16 @@
 import { withCurrentProduct } from '@shopgate/engage/core';
 import React from 'react';
 import PropTypes from 'prop-types';
+import { css } from 'glamor';
 import I18n from '@shopgate/pwa-common/components/I18n';
 import { ITEM_PATH } from '@shopgate/pwa-common-commerce/product/constants';
 import { bin2hex } from '@shopgate/pwa-common/helpers/data';
 import ButtonLink from '@shopgate/pwa-ui-shared/ButtonLink';
 import connect from './connector';
+
+const buttonStyle = css({
+  paddingRight: 0,
+}).toString();
 
 /**
  * Link to add a review.
@@ -20,7 +25,7 @@ const WriteReviewLink = ({ productId }) => {
 
   return (
     <div data-test-id="writeReview">
-      <ButtonLink href={`${ITEM_PATH}/${publicProductId}/write_review`} noGap>
+      <ButtonLink className={buttonStyle} href={`${ITEM_PATH}/${publicProductId}/write_review`} noGap>
         <I18n.Text string="reviews.button_add" />
       </ButtonLink>
     </div>