+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ### Added
 - Classes for container, section, section header and section content
+- New accordion item type `properties` to move the whole properties section into an accordion
 - `sectionBorderColor` config parameter to change the color of the section borders
 - `showLastSectionBottomBorder` config parameter to show a border at the bottom of the last section
 - `sectionButtonIcon` config parameter to replace the default section SVG icon (will be turned by 180° in open state by default)